Timer driver updated

+++ utils/stm32f411ret6/delay.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 void delay_init();
-void delay_ms(uint32_t t);
-void delay_us(uint32_t t);
+void delay_ms(uint16_t t);
+void delay_us(uint16_t t);
 
 #endif
